school/requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
Faker==28.4.1
Markdown==3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
validate-docbr==1.10.0
